Move human readable size to napari utils misc
